Add a notification message to prevent error (#19)
Add a notification message to prevent error

Without the notification message, when grep cannot find that a branch is
ahead (i.e. when nothing is wrong), the exit code of the grep command is
1, and this can propagate out of the loop and exit the entire biowdl
script with an error.

Since we call exit 1 explicitly when we encounter a branch that is
ahead, adding this echo statement will make sure that the return code is
0 when no branches have been found to be ahead.

Remove the subshell for exiting when a branch is ahead, since that doesn't
propagate the error to the outer script properly (due to the echo statement
being executed).